feat: Add endpoint to add time to timer
This commit adds a new endpoint `add_time` to the `TimerViewSet` in the `views.py` file. The endpoint allows authenticated users to add time to the timer by specifying the amount of time in seconds. The `add_time` method in the `Timer` model is called to update the `timer_end` attribute and save the changes. After adding the time, the `send_ticket` method is called to send a ticket. If the time provided is invalid, a response with status code 400 and a message indicating the invalid time is returned.

Note: This commit message follows the conventional format of starting with a type (feat) followed by a brief description of the changes.